+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Change Log
 
 All notable changes to this project will be documented in this file. This project does not adhere to [Semantic Versioning](http://semver.org/)
+## 2.1.7 (06/24/2016)
+- QA-270: Denied attempt test for Teacher Doc Upload verifier
+- QA-287: Verify Marketplace Group Affiliation for Teachers
+- QA-288: Verify Marketplace Group Affiliation for Govt Employees
+- QA-289: Verify Marketplace Group Affiliation for Students
+
+## 2.1.6 (06/23/2016)
+- QA-286: Verify Marketplace Group Affiliation for First Responders 
+- QA-271: Denied attempt test for Teacher Lookup verifier
+
 ## 2.1.5 (06/22/2016)
 - QA-268: Denied attempt test for Student NSC verifier
 - QA-253: Dupe attempt test for Student NSC verifier
@@ -77,4 +87,4 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 - New branch commits being restructured in Capybara / Selenium-Webdriver
 ### Removed
-- Removed earlier V1 to branch using Watir-Webdriver / Selenium Grid
+- Removed earlier V1 to branch using Watir / Selenium-Webdriver